DAOArquivo

+++ RikStokPro/src/EDA/Secos.java
@@ -2,14 +2,14 @@ package EDA;
 
 public class Secos extends Produto {
 
-    private int peso;
+    private Double peso;
 
-    public Secos(int peso, String nome, int id, int quantidade, double valor) {
+    public Secos(Double peso, String nome, int id, int quantidade, double valor) {
         super(nome, id, quantidade, valor);
         this.peso = peso;
     }
 
-    public int getPeso() {
+    public Double getPeso() {
         return peso;
     }
